Move Network to core
