+++ core/src/network.rs
@@ -1,4 +1,5 @@
 use async_trait::async_trait;
+use clap::ValueEnum;
 use color_eyre::{eyre::WrapErr, Result};
 use dusk_plonk::prelude::PublicParameters;
 use kate_recovery::{
@@ -9,6 +10,7 @@ use kate_recovery::{
 use mockall::automock;
 use sp_core::H256;
 use std::{sync::Arc, time::Duration};
+use strum::Display;
 use tokio::time::Instant;
 use tracing::{debug, info};
 
@@ -215,5 +217,84 @@ pub fn new(
 		rpc_client,
 		pp,
 		disable_rpc,
+	}
+}
+
+#[derive(Clone, ValueEnum, Display)]
+#[strum(serialize_all = "kebab-case")]
+pub enum Network {
+	Local,
+	Hex,
+	Turing,
+	Mainnet,
+}
+
+impl Network {
+	pub fn bootstrap_peer_id(&self) -> &str {
+		match self {
+			Network::Local => "12D3KooWStAKPADXqJ7cngPYXd2mSANpdgh1xQ34aouufHA2xShz",
+			Network::Hex => "12D3KooWBMwfo5qyoLQDRat86kFcGAiJ2yxKM63rXHMw2rDuNZMA",
+			Network::Turing => "12D3KooWBkLsNGaD3SpMaRWtAmWVuiZg1afdNSPbtJ8M8r9ArGRT",
+			Network::Mainnet => "12D3KooW9x9qnoXhkHAjdNFu92kMvBRSiFBMAoC5NnifgzXjsuiM",
+		}
+	}
+
+	pub fn bootstrap_multiaddrr(&self) -> &str {
+		match self {
+			Network::Local => "/ip4/127.0.0.1/tcp/39000",
+			Network::Hex => "/dns/bootnode.1.lightclient.hex.avail.so/tcp/37000",
+			Network::Turing => "/dns/bootnode.1.lightclient.turing.avail.so/tcp/37000",
+			Network::Mainnet => "/dns/bootnode.1.lightclient.mainnet.avail.so/tcp/37000",
+		}
+	}
+
+	pub fn full_node_ws(&self) -> Vec<String> {
+		match self {
+			Network::Local => vec!["ws://127.0.0.1:9944".to_string()],
+			Network::Hex => vec!["wss://rpc-hex-devnet.avail.tools/ws".to_string()],
+			Network::Turing => vec!["wss://turing-rpc.avail.so/ws".to_string()],
+			Network::Mainnet => vec![
+				"wss://mainnet.avail-rpc.com/".to_string(),
+				"wss://avail-mainnet.public.blastapi.io/".to_string(),
+				"wss://mainnet-rpc.avail.so/ws".to_string(),
+			],
+		}
+	}
+
+	pub fn ot_collector_endpoint(&self) -> &str {
+		match self {
+			Network::Local => "http://127.0.0.1:4317",
+			Network::Hex => "http://otel.lightclient.hex.avail.so:4317",
+			Network::Turing => "http://otel.lightclient.turing.avail.so:4317",
+			Network::Mainnet => "http://otel.lightclient.mainnet.avail.so:4317",
+		}
+	}
+
+	pub fn genesis_hash(&self) -> &str {
+		match self {
+			Network::Local => "DEV",
+			Network::Hex => "9d5ea6a5d7631e13028b684a1a0078e3970caa78bd677eaecaf2160304f174fb",
+			Network::Turing => "d3d2f3a3495dc597434a99d7d449ebad6616db45e4e4f178f31cc6fa14378b70",
+			Network::Mainnet => "b91746b45e0346cc2f815a520b9c6cb4d5c0902af848db0a80f85932d2e8276a",
+		}
+	}
+
+	pub fn name(genesis_hash: &str) -> String {
+		let network = match genesis_hash {
+			"b91746b45e0346cc2f815a520b9c6cb4d5c0902af848db0a80f85932d2e8276a" => {
+				Network::Mainnet.to_string()
+			},
+			"9d5ea6a5d7631e13028b684a1a0078e3970caa78bd677eaecaf2160304f174fb" => {
+				Network::Hex.to_string()
+			},
+			"d3d2f3a3495dc597434a99d7d449ebad6616db45e4e4f178f31cc6fa14378b70" => {
+				Network::Turing.to_string()
+			},
+			"DEV" => Network::Local.to_string(),
+			_ => "other".to_string(),
+		};
+
+		let prefix = &genesis_hash[..std::cmp::min(6, genesis_hash.len())];
+		format!("{}:{}", network, prefix)
 	}
 }
